+++ src/common/Errors.java
@@ -361,6 +361,13 @@ public enum Errors {
             ============================================================
             [DB] 실사로그를 작성한 창고관리자를 확인하는 중 문제가 생겼습니다.
             ============================================================
+            """),
+
+
+    WA_HAVE_NO_WIDX("""
+            ============================================================
+            현재 창고관리자는 창고 배정을 받지 않았습니다.
+            ============================================================
             """);
